[CD-153] Move: subdivide create_user_script folder
