+++ front/lib/screen/script/create_user_script/input_field/category_section.dart
